backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.4",
    "fastapi[standard]>=0.133.1",
    "pwdlib[argon2]>=0.3.0",
    "pyjwt>=2.11.0",
    "pytest>=9.0.2",
    "sqlmodel>=0.0.37",
]
